Improve package name for ck
