+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/common/jooq/tables/BithonTraceSpanSummary.java
@@ -4,10 +4,6 @@
 package org.bithon.server.storage.jdbc.common.jooq.tables;
 
 
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.List;
-
 import org.bithon.server.storage.jdbc.common.jooq.DefaultSchema;
 import org.bithon.server.storage.jdbc.common.jooq.Indexes;
 import org.bithon.server.storage.jdbc.common.jooq.tables.records.BithonTraceSpanSummaryRecord;
@@ -22,6 +18,10 @@ import org.jooq.Table;
 import org.jooq.TableField;
 import org.jooq.impl.DSL;
 import org.jooq.impl.TableImpl;
+
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
 
 
 /**
