+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/tracing/writer/MappingTableWriter.java
@@ -17,6 +17,7 @@
 package org.bithon.server.storage.jdbc.tracing.writer;
 
 import org.bithon.component.commons.utils.RetryUtils;
+import org.bithon.server.storage.jdbc.common.IOnceTableWriter;
 import org.bithon.server.storage.jdbc.common.jooq.Tables;
 import org.bithon.server.storage.tracing.mapping.TraceIdMapping;
 import org.jooq.DSLContext;
@@ -32,7 +33,7 @@ import java.util.function.Predicate;
  * @author Frank Chen
  * @date 16/1/24 8:01 pm
  */
-class MappingTableWriter implements ITableWriter {
+class MappingTableWriter implements IOnceTableWriter {
     private final String insertStatement;
     private final Collection<TraceIdMapping> mappings;
     private final Predicate<Exception> isExceptionRetryable;
